A finite cache fix and a related validator fix, plus some documentation.
1) The finite cache was completely broken.  In particular, the way it kept track of
what to evict from the cache next relied on false assumptions about the pointers
the validator handed it.  The finite cache assumed that the pointers were
basically const, but the validator reused the memory.

2) The validator was not correctly checking whether each policy result should be
cached.  This causes problems for policies like heap, where some rules must not
be cached.  When using heap with a sufficiently large cache, this bug would
result in the use of a single color everywhere, because the rule that was
supposed to generate new colors each time it fired got cached and only fired
once.