Only parse the ASM line up to the :
For some reason, on longjump_works_1/testSimple, using strtoul on the
whole line or any character past the address (including the :) causes a
segfault. strotul is supposed to be able to handle that and even
includes functionality for retrieving the index of the first character
not parsed, so it's not clear why this is happening.