Make meta set cache work with mutable sets
The generated policy evaluation code directly modifies a meta
set that has been stored in the meta set cache.  This makes the former
scheme of mapping meta sets onto "canon" global versions a potential
source of problems, as after modifying the canon version the key of the
map no longer matches the value. This can result in duplicate entries
caused by "canonizing" newly-modified meta sets or getting the wrong
meta set when attempting to retrive the canon version of a set that has
been modified using the unmodified version.  In order to accomodate
this, the map is replaced with a list so the meta sets are no longer
associated with keys and can be freely modified. This has a side effect
of potentially slowing down the program, but unless the number of meta
sets or lookups is very large it shouldn't be noticeable.

This should also help pave the way for changing tags from being pointers
converted to integers to being something else thats less dangerous and
supports tags on unaligned data (e.g. for compressed instructions).