Reserve large initial capacity for meta set vector
Some of the generated policy code copies meta set pointers into the
individual tags of other meta sets and then later uses the data they
point to. When std::vector grows beyond its capacity, it copies the
entire array to a new location with a new capacity to fit the new data,
which would invalidate these pointers that have been copied. A warning
is also printed to indicate that this has happened and it could affect
future policy execution.